compounds: Try struct-like definitions before lex rules
Giving the 'type' rule a higher priority than to 'lex' is necessary to
let statements like the following be parsed properly:

    struct /* comment */ name {
        ...
    };

Otherwise, at the point just after the 'struct' keyword, the 'comments'
rule is taken instead of the proper 'type-definition' one. This is
because the latter uses positive lookbehind to match the keyword, i.e.
(?<= enum | class | union | struct ), that can't trigger once the
comment is consumed.

Related changeset:

    da66703 ("fix broken handling of typedef struct ... alias")