compounds: Fix corner-case that might break func detection heuristics
The 'type-definition' rule in question uses a positive lookahead in its
'end' pattern to check whether the 'function-definition' rule can be
taken next (the pattern is almost identical to that one from 'begin' of
the 'function-definition' rule).

That pattern uses lookbehind groups to make sure the context is sane,
before matching a function name. However, these lookbehind groups were
broken because of using '\s*' just in front of them.