compounds: Recognize struct name even if no open brace follows it
This fixes broken symbol indexing and "Goto Definition" for code that
follows a style guide advising to put open braces at the beginning
of the next line, after the line containing the struct name. For
example:

    struct name  // <- was absent from the symbol index until now
    {
        ...
    };

However, there are false positives in certain (rare) cases, like the
following:

    struct point  // <- will be mistakenly recognized as a struct definition
    create_point(int x, int y);

Nevertheless, I believe, these cases are uncommon: some people really
do tend to put the return type of a function to a distinct line, but
structures are usually returned by a pointer, not by value, and
enumerations are not so common per se.

That is, these cases are much less frequent than real and valid struct
definitions across two lines.

Suggested by @jaytiar. Closes #24.